Rework the db model 'Administrator' and relocate constants for the factories
