+++ src/admin_user/factories/factories.py
@@ -2,37 +2,37 @@ import factory
 from factory.django import DjangoModelFactory
 from factory.fuzzy import FuzzyInteger
 
-from admin_user.constants import (
+from admin_user.models import Administrator
+
+from .constants import (
     ADMIN_CREATION_INSTANCES_COUNT,
     END_PHONE_NUMBER_VALUE,
     START_PHONE_NUMBER_VALUE,
 )
-from admin_user.models import Administrator
 
 
 class AdminFactory(DjangoModelFactory):
-    """Factory of Admin for the project testing."""
+    """Factory of `Administrator` for the project testing."""
 
     class Meta:
         model = Administrator
 
-    first_name = factory.Faker("first_name")
-    last_name = factory.Faker("last_name")
+    first_name = factory.Faker('first_name')
+    last_name = factory.Faker('last_name')
     phone = factory.Sequence(
-        lambda some: f"+7495{FuzzyInteger(
+        lambda some: f'+7495{FuzzyInteger(
             START_PHONE_NUMBER_VALUE,
             END_PHONE_NUMBER_VALUE,
-        ).fuzz()}",
+        ).fuzz()}',
     )
     email = factory.LazyAttribute(
-        lambda a: f"{a.first_name}.{a.last_name}@example.com".lower(),
+        lambda a: f'{a.first_name}.{a.last_name}@example.com'.lower(),
     )
-    username = factory.Faker("user_name", locale="ru_RU")
 
     @classmethod
     def _create(cls, model_class, *args, **kwargs):
-        """Override default _create method to set password."""
-        password = kwargs.pop("password", None)
+        """Override default `_create` method to set password."""
+        password = kwargs.pop('password', None)
         administrator = super()._create(model_class, *args, **kwargs)
         administrator.set_password(password)
         administrator.is_staff = True
@@ -41,5 +41,5 @@ class AdminFactory(DjangoModelFactory):
 
 
 def create_admins():
-    """Create Administrator instances for the project tests."""
+    """Create `Administrator` instances for the project tests."""
     AdminFactory.create_batch(size=ADMIN_CREATION_INSTANCES_COUNT)
