+++ src/admin_user/models.py
@@ -9,17 +9,18 @@ EMAIL_LENGTH = 254
 class Administrator(AbstractUser):
     """Custom Administrator model."""
 
+    username = None
     first_name = models.CharField('Имя', max_length=DEFAULT_NAME_LENGTH)
     last_name = models.CharField('Фамилия', max_length=DEFAULT_NAME_LENGTH)
     phone = PhoneNumberField('Номер телефона', help_text='Формат +7XXXXXXXXXX')
-    USERNAME_FIELD = 'email'
-    REQUIRED_FIELDS = ('username', 'first_name', 'last_name', 'phone')
-
     email = models.EmailField(
         'E-mail адрес',
         max_length=EMAIL_LENGTH,
         unique=True,
     )
+
+    USERNAME_FIELD = 'email'
+    REQUIRED_FIELDS = ('first_name', 'last_name', 'phone')
 
     class Meta:
         """
@@ -34,4 +35,4 @@ class Administrator(AbstractUser):
 
     def __str__(self) -> str:
         """Return a string representation of User object."""
-        return self.username
+        return f'{self.first_name} {self.last_name}'
