+++ src/core/config/settings_base.py
@@ -17,7 +17,7 @@ DEFAULT = 'some_default_key'
 
 SECRET_KEY = env.str('SECRET_KEY', default=DEFAULT)
 
-TELEGRAM_TOKEN = env.str("TELEGRAM_TOKEN")
+TELEGRAM_TOKEN = env.str('TELEGRAM_TOKEN')
 
 ALLOWED_HOSTS = ['*']
 
@@ -37,7 +37,7 @@ LOCAL_APPS = [
     'schooling.apps.SchoolingConfig',
 ]
 
-EXTERNAL_APPS = ["phonenumber_field"]
+EXTERNAL_APPS = ['phonenumber_field']
 
 INSTALLED_APPS = DEFAULT_APPS + EXTERNAL_APPS + LOCAL_APPS
 
@@ -100,4 +100,4 @@ STATIC_URL = 'static/'
 
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
-PHONENUMBER_DEFAULT_REGION = "RU"
+PHONENUMBER_DEFAULT_REGION = 'RU'
